+++ projects/templates/src/lib/components/po-page-dynamic-search/index.ts
@@ -1,5 +1,6 @@
 export * from './po-page-dynamic-search.component';
 export * from './po-page-dynamic-search.interface';
 export * from './po-page-dynamic-search-literals.interface';
+export * from './po-page-dynamic-search-options.interface';
 
 export * from './po-page-dynamic-search.module';