+++ projects/templates/src/lib/components/po-page-dynamic-table/index.ts
@@ -1,5 +1,7 @@
 export * from './po-page-dynamic-table.component';
-export * from './po-page-dynamic-table-actions.interface';
-export * from './po-page-dynamic-table-field.interface';
+export * from './interfaces/po-page-dynamic-table-actions.interface';
+export * from './interfaces/po-page-dynamic-table-field.interface';
+export * from './interfaces/po-page-dynamic-table-options.interface';
+export * from './interfaces/po-page-dynamic-table-metadata.interface';
 
 export * from './po-page-dynamic-table.module';