+++ projects/templates/src/lib/components/po-page-dynamic-table/po-page-dynamic-list-base.component.ts
@@ -4,7 +4,7 @@ import { PoBreadcrumb } from '@portinari/portinari-ui';
 
 import { convertToBoolean } from '../../utils/util';
 
-import { PoPageDynamicTableField } from './po-page-dynamic-table-field.interface';
+import { PoPageDynamicTableField } from './interfaces/po-page-dynamic-table-field.interface';
 
 export class PoPageDynamicListBaseComponent {