+++ projects/templates/src/lib/services/po-page-dynamic/index.ts
@@ -1,0 +1,3 @@
+export * from './po-page-dynamic.service';
+
+export * from './po-page-dynamic.module';